split out an inner class
